LayoutTests/http/tests/media/video-served-as-text-expected.txt

EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(loadedmetadata)
loaded media file served as text OK
END OF TEST
